+++ .vs/VSWorkspaceState.json
@@ -7,6 +7,6 @@
     "\\src\\js",
     "\\src\\product_pages"
   ],
-  "SelectedNode": "\\src\\js\\product.js",
+  "SelectedNode": "\\src\\js\\shake.js",
   "PreviewInSolutionExplorer": false
 }
