+++ src/js/checkout.js
@@ -1,0 +1,2 @@
+import { loadHeaderFooter } from "./utils.mjs"
+loadHeaderFooter();
