+++ src/js/product.js
@@ -1,5 +1,8 @@
 import { getParam } from "./utils.mjs";
 import productDetails from "./productDetails.mjs";
+import { loadHeaderFooter } from "./utils.mjs"
+loadHeaderFooter();
+
 
 const productId = getParam("product");
 productDetails(productId);
